Resolved conflicts and added switcher.py
